Initial commit, 2021-07-10
No, there is no history going back, because I made a complete mess on trying to store the true "source" for my website, which containts large resolution images, and a few more things in addition.

I finally came to realize that what I actually need in a git repository is my end result website, hence this one here, but the amount of work required to export the "history" of a file from the other repository is just insane, so no history.